pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "troer"
version = "0.1.0"
description = "IDL translatore to dpack serializer"
urls.Homepage = "https://github.com/Geneo-5/troer"
urls.Issues = "https://github.com/Geneo-5/troer/issues"
urls.Repository = "https://github.com/Geneo-5/troer.git"
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
	"pyyaml",
	"jsonschema>=4.23",
	"Cheetah3",
]
scripts.troer = "troer.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/troer"]
artifacts = [
	"src/troer/schema/*.yaml",
	"src/troer/templates/*.tmpl"
]
